+++ CHANGELOG.md
@@ -9,4 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - GLPI 12 Compatibility
 
+### Fixed
+- Fixed the issue where a ticket could be solved without a solution
+
 ## [1.0.0-rc1]